refactor : hook 구조 일부 변경
