+++ src/pages/apply/Apply.tsx
@@ -6,15 +6,15 @@ import { z } from 'zod';
 import { useParams } from 'react-router-dom';
 import { formatDate } from '../../util/format';
 import { joinProject } from '../../models/joinProject';
-import useGetProjectData from '../../hooks/useJoinProject';
 import CareersComponent from '../../components/applyComponents/careersComponent/CareersComponent';
 import PhoneComponent from '../../components/applyComponents/phoneComponent/PhoneComponent';
 import LoadingSpinner from '../../components/common/loadingSpinner/LoadingSpinner';
 import Modal from '../../components/common/modal/Modal';
 import { useModal } from '../../hooks/useModal';
-import useApplyProject from '../../hooks/useApplyProject';
 import useAuthStore from '../../store/authStore';
 import { useEffect } from 'react';
+import useGetProjectData from '../../hooks/useGetProjectData';
+import useApplyProject from '../../hooks/ProjectHooks/useApplyProject';
 
 const ApplyScheme = z.object({
   email: z
